Move image analysis in separate file.
