--- sandbox.py
@@ -1,23 +1,0 @@
-# Collections
-
-
-
-# List
-test_list = [1, 2, [], 0.7, -6]
-
-test_list.append('new')
-test_list.
-
-print(test_list)
-
-
-test_dict = {"Sasha":"The best", 8:"Fox"}
-
-
-test_dict["new"] = 3
-
-print(type(test_dict))
-print(test_dict)
-
-test_dict["Sasha"] = "Dragon"
-print(test_dict["Sasha"])
